requirements.txt
streamlit==1.31.1
pandas==2.2.0
Pillow 